Move input-number to a common directory
