+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill.js
@@ -7,7 +7,7 @@ var DialogView = require('builder/components/dialog/dialog-view');
 var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 
 var InputCollection = require('./input-collection');
-var InputNumber = require('./input-number/input-number');
+var InputNumber = require('builder/components/input-number/input-number');
 var InputColor = require('builder/components/input-color/input-color');
 
 var template = require('./fill-template.tpl');
