+++ lib/assets/javascripts/builder/components/input-number/input-number-content-view.tpl
