+++ lib/assets/javascripts/builder/components/input-number/input-number-fixed-content-view.js
