+++ lib/assets/javascripts/builder/components/input-number/input-number-value-content-view.tpl
