+++ lib/assets/javascripts/builder/components/input-number/input-number.tpl
