+++ lib/assets/javascripts/builder/components/input-number/quantification-method-item.tpl
