+++ lib/assets/test/spec/builder/components/form-components/editors/fill/input-number.spec.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var Backbone = require('backbone');
-var InputNumber = require('builder/components/form-components/editors/fill/input-number/input-number');
+var InputNumber = require('builder/components/input-number/input-number');
 
 describe('components/form-components/editors/fill/input-number', function () {
   beforeEach(function () {
