+++ lib/assets/test/spec/builder/components/form-components/editors/fill/input-number/input-number-content-view.spec.js
@@ -1,6 +1,6 @@
 var cdb = require('internal-carto.js');
 var Backbone = require('backbone');
-var InputNumberContentView = require('builder/components/form-components/editors/fill/input-number/input-number-content-view');
+var InputNumberContentView = require('builder/components/input-number/input-number-content-view');
 
 describe('components/form-components/editors/fill/input-number/input-number-content-view', function () {
   describe('on model with a range', function () {
