+++ lib/assets/test/spec/builder/components/form-components/editors/fill/input-number/input-number-dialog-content.spec.js
@@ -1,5 +1,5 @@
 var Backbone = require('backbone');
-var InputNumberDialogContent = require('builder/components/form-components/editors/fill/input-number/input-number-dialog-content');
+var InputNumberDialogContent = require('builder/components/input-number/input-number-dialog-content');
 
 var COLUMNS = ['schmeckles', 'flurbos'];
