+++ lib/assets/test/spec/builder/components/form-components/editors/fill/input-number/input-number-value-content-view.spec.js
@@ -1,5 +1,5 @@
 var Backbone = require('backbone');
-var InputNumberValueContentView = require('builder/components/form-components/editors/fill/input-number/input-number-value-content-view');
+var InputNumberValueContentView = require('builder/components/input-number/input-number-value-content-view');
 
 describe('components/form-components/editors/fill/input-number/input-number-value-content-view', function () {
   describe('on model with a range', function () {
